Refactor monolithic script.js into modules
Decomposes the large `script.js` file into smaller, more manageable modules based on functionality.

Key changes:
- Created a new `js/modules` directory to house the new modules, including `auth.js`, `api.js`, `ui.js`, `calendar.js`, and `ai.js`.
- The new `js/app.js` now serves as the main application entry point, importing from the other modules to orchestrate the application.
- Updated `index.html` to load `js/app.js` with `type="module"`.
- Added a `.gitignore` file to exclude `node_modules` from version control.
- Added `netlify-cli` as a development dependency in `package.json` to support local testing and verification.

This refactoring significantly improves the project structure and makes the codebase easier to understand and maintain in the future.
