+++ index.html
@@ -395,7 +395,7 @@
     <script src="https://www.gstatic.com/firebasejs/9.6.7/firebase-auth-compat.js"></script>
     <script src="https://www.gstatic.com/firebasejs/9.6.7/firebase-firestore-compat.js"></script>
 
-    <script src="script.js"></script>
+    <script type="module" src="js/app.js"></script>
 
 </body>
 </html>
