+++ js/modules/dom.js
@@ -1,0 +1,59 @@
+export const DOMElements = {
+    initialLoadingView: document.getElementById('initial-loading-view'),
+    loadingView: document.getElementById('loading-view'),
+    creationLoadingView: document.getElementById('creation-loading-view'),
+    loginView: document.getElementById('login-view'),
+    dashboardView: document.getElementById('dashboard-view'),
+    dashboardContent: document.getElementById('dashboard-content'),
+    appView: document.getElementById('app-view'),
+    mainContent: document.querySelector('#app-view main'),
+    sidebar: document.getElementById('sidebar'),
+    emailInput: document.getElementById('email'),
+    passwordInput: document.getElementById('password'),
+    loginBtn: document.getElementById('login-btn'),
+    sidebarLogoutBtn: document.getElementById('sidebar-logout-btn'), // Corrected reference
+    dashboardLogoutBtn: document.getElementById('dashboard-logout-btn'),
+    dashboardProfileBtn: document.getElementById('dashboard-profile-btn'),
+    backToDashboardBtn: document.getElementById('back-to-dashboard-btn'),
+    authError: document.getElementById('auth-error'),
+    sidebarName: document.getElementById('sidebar-name'),
+    sidebarBakery: document.getElementById('sidebar-bakery'),
+    sidebarInitials: document.getElementById('sidebar-initials'),
+    contentArea: document.getElementById('content-area'),
+    mainNav: document.getElementById('main-nav'),
+    printBtn: document.getElementById('print-btn'),
+    shareBtn: document.getElementById('share-btn'),
+    saveIndicator: document.getElementById('save-indicator'),
+    headerTitle: document.getElementById('header-title'),
+    headerSubtitle: document.getElementById('header-subtitle'),
+    progressBarFill: document.getElementById('progress-bar-fill'),
+    progressPercentage: document.getElementById('progress-percentage'),
+    desktopHeaderButtons: document.getElementById('desktop-header-buttons'),
+    aiActionBtn: document.getElementById('ai-action-btn'),
+    mobileMenuBtn: document.getElementById('mobile-menu-btn'),
+    sidebarOverlay: document.getElementById('sidebar-overlay'),
+    modalOverlay: document.getElementById('modal-overlay'),
+    modalBox: document.getElementById('modal-box'),
+    modalTitle: document.getElementById('modal-title'),
+    modalContent: document.getElementById('modal-content'),
+    modalActionBtn: document.getElementById('modal-action-btn'),
+    modalCancelBtn: document.getElementById('modal-cancel-btn'),
+    modalCloseBtn: document.getElementById('modal-close-btn'),
+    resetView: document.getElementById('reset-view'),
+    forgotPasswordBtn: document.getElementById('forgot-password-btn'),
+    resetEmail: document.getElementById('reset-email'),
+    sendResetBtn: document.getElementById('send-reset-btn'),
+    resetMessageContainer: document.getElementById('reset-message-container'),
+    backToLoginBtn: document.getElementById('back-to-login-btn'),
+    registerView: document.getElementById('register-view'),
+    showRegisterViewBtn: document.getElementById('show-register-view-btn'),
+    backToLoginFromRegisterBtn: document.getElementById('back-to-login-from-register-btn'),
+    registerEmail: document.getElementById('register-email'),
+    registerPassword: document.getElementById('register-password'),
+    termsAgreeCheckbox: document.getElementById('terms-agree'),
+    createAccountBtn: document.getElementById('create-account-btn'),
+    registerError: document.getElementById('register-error'),
+    radialMenuContainer: document.getElementById('radial-menu-container'),
+    radialMenuFab: document.getElementById('radial-menu-fab'),
+    radialMenuOverlay: document.getElementById('radial-menu-overlay'),
+};
