+++ package.json
@@ -9,6 +9,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@google/generative-ai": "^0.11.3"
+    "@google/generative-ai": "^0.11.3",
+    "netlify-cli": "^23.4.3"
   }
 }
